adding checkbox method, clear method and array reaarange
