+++ src/index.html
@@ -33,7 +33,7 @@
                 </div>
             </div>
             <div class="button-div">
-                <button id="clear" type="submit">Clear All Completed </button>
+                <button id="clear">Clear All Completed </button>
             </div>
         </form>
         </section>
